+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## 0.3.0
+# 0.4.0
+- Added a timeout to the Mailbox.take, Mutex.runLocked and ConditionVariable.wait methods.
+ Note: the Mutex timeout is ignored on Windows.
 
+
+## 0.3.0
 - Add a closed state to `Mailbox`.
 
 ## 0.2.0